requirements.txt

configparser==3.5.0
decorator==4.0.9
Django==1.9
django-debugtools==1.6
django-mysql==1.0.9
greenlet==0.4.9
ipython==4.1.2
ipython-genutils==0.1.0
paramiko==1.16.0
path.py==8.1.2
pexpect==4.0.1
PyMySQL==0.7.3
PyYAML==3.11

